FEAT : Rename package from common to global common 대신 global이라는 패키로 이름 변경
