+++ src/main/java/com/ssg/starroadadmin/board/entity/BoardImage.java
@@ -1,6 +1,6 @@
 package com.ssg.starroadadmin.board.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
