+++ src/main/java/com/ssg/starroadadmin/coupon/entity/Coupon.java
@@ -1,6 +1,6 @@
 package com.ssg.starroadadmin.coupon.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
