+++ src/main/java/com/ssg/starroadadmin/review/entity/ReviewImage.java
@@ -1,6 +1,6 @@
 package com.ssg.starroadadmin.review.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
