+++ src/main/java/com/ssg/starroadadmin/reward/entity/Reward.java
@@ -1,6 +1,6 @@
 package com.ssg.starroadadmin.reward.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
