+++ src/main/java/com/ssg/starroadadmin/reward/entity/RewardHistory.java
@@ -1,6 +1,6 @@
 package com.ssg.starroadadmin.reward.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
 import com.ssg.starroadadmin.user.entity.Users;
 import jakarta.persistence.*;
 import lombok.Getter;
