+++ src/main/java/com/ssg/starroadadmin/shop/entity/Store.java
@@ -1,8 +1,10 @@
 package com.ssg.starroadadmin.shop.entity;
 
-import com.ssg.starroadadmin.common.entity.BaseTimeEntity;
-import com.ssg.starroadadmin.user.entity.Users;
+import com.ssg.starroadadmin.global.entity.BaseTimeEntity;
+import com.ssg.starroadadmin.user.entity.Manager;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,7 +13,9 @@ import static lombok.AccessLevel.PROTECTED;
 
 @Getter
 @Entity
+@Builder
 @NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor
 public class Store extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -23,7 +27,7 @@ public class Store extends BaseTimeEntity {
 
     @ManyToOne
     @JoinColumn(name = "manager_id")
-    private Users managerId;
+    private Manager manager;
 
     @Column(unique = true)
     private String name;
@@ -33,4 +37,18 @@ public class Store extends BaseTimeEntity {
     private int floor;
     private String operatingTime; // hh:mm ~ hh:mm format
     private String contactNumber;
+
+    public void updateInfo(String imagePath, String contents, int floor, String operatingTime, String contactNumber) {
+        this.imagePath = checkNull(imagePath, this.imagePath);
+        this.contents = checkNull(contents, this.contents);
+        this.floor = checkNull(floor, this.floor);
+        this.operatingTime = checkNull(operatingTime, this.operatingTime);
+        this.contactNumber = checkNull(contactNumber, this.contactNumber);
+    }
+
+    private <T> T checkNull(T newValue, T currentValue) {
+        return newValue != null ? newValue : currentValue;
+    }
+
+
 }
